pyproject.toml
[tool.poetry]
name = "telegram_mailing_help"
version = "1.1.0"
description = ""
authors = ["cherkashin <alexander.cherkashin@sibit.tech>"]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^13.1"
fastapi = ">=0.110.0"
uvicorn = ">=0.29.0"
python-multipart = ">=0.0.9"
yoyo-migrations = "^7.2.1"
dacite = "^1.6.0"
sqlite3worker = "^1.1.7"
gevent = "^22.10.2"
tabulate = "^0.8.7"
requests = "^2.26.0"
PyJWT = ">=2.8.0"
pywebpush = ">=2.0.0"
setuptools = ">=65.0.0,<71"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-playwright = "^0.4"
testcontainers = {extras = ["selenium"], version = "^3.4.2"}
responses = ">=0.23"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
